♻️ Refactor: 스티커 목록 공통 컴포넌트로 이동
