+++ apps/web/src/shared/ui/sticker-collection/StickerCollection.tsx
@@ -6,14 +6,16 @@ type StickerCollectionProps = {
   ariaLabel?: string;
   className?: string;
   maxItems?: number;
+  size?: 'default' | 'compact';
   stickers: readonly string[];
 };
 
-/** 지도 장소 시트에서 스티커를 5열 슬롯으로 표시합니다. */
+/** 스티커를 5열 슬롯에 표시하고 마지막 줄의 남은 자리를 빈 슬롯으로 채웁니다. */
 export default function StickerCollection({
   ariaLabel = '스티커 목록',
   className,
   maxItems,
+  size = 'default',
   stickers,
 }: StickerCollectionProps) {
   const visibleStickers = maxItems === undefined ? stickers : stickers.slice(0, maxItems);
@@ -25,23 +27,35 @@ export default function StickerCollection({
     { length: slotCount },
     (_, index) => visibleStickers[index] ?? null
   );
+  const isCompact = size === 'compact';
 
   return (
     <ul
       aria-label={ariaLabel}
       className={cn(
-        'grid grid-cols-5 items-center gap-y-4 rounded-16 bg-neutral-50 p-4',
+        'grid grid-cols-5 items-center rounded-16 bg-neutral-50 py-4',
+        isCompact ? 'gap-x-4 gap-y-8 px-2.75' : 'gap-y-4 px-4',
         className
       )}
     >
       {stickerSlots.map((stickerImage, index) => (
-        <li key={`sticker-slot-${index}`} className="flex w-18 justify-self-center justify-center">
+        <li
+          key={`sticker-slot-${index}`}
+          className={cn('flex justify-self-center justify-center', isCompact ? 'w-13.75' : 'w-18')}
+        >
           {stickerImage ? (
-            <img src={stickerImage} alt="" className="size-18 object-contain" />
+            <img
+              src={stickerImage}
+              alt=""
+              className={cn('object-contain', isCompact ? 'size-13.75' : 'size-18')}
+            />
           ) : (
             <span
               aria-label="빈 스티커 자리"
-              className="block size-12 rounded-full border border-dashed border-neutral-400 bg-neutral-100"
+              className={cn(
+                'block rounded-full border border-dashed border-neutral-400 bg-neutral-100',
+                isCompact ? 'size-11.25' : 'size-12'
+              )}
             />
           )}
         </li>
