+++ apps/web/src/shared/ui/sticker-collection/index.ts
@@ -1,0 +1,1 @@
+export { default as StickerCollection } from './StickerCollection';
